+++ packages/mongo-async/collection_tests.js
@@ -147,8 +147,8 @@ Tinytest.addAsync('collection - calling native find with good hint and maxTimeMs
     Promise.resolve(
       Meteor.isServer &&
         collection.rawCollection().createIndex({ a: 1 })
-    ).then(() => {
-      test.equal(collection.find({}, {
+    ).then(async () => {
+      test.equal(await collection.find({}, {
         hint: {a: 1},
         maxTimeMs: 1000
       }).count(), 1);